Complete exercise 4.21*

+++ controllers/blogs.js
@@ -50,6 +50,18 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (!(blog.user.toString() === user._id.toString())) {
+    return response.status(403).json({ error: 'action forbidden' });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
